refactor: Convert useSharedDriveFolder to tsx
